+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Release Notes
 
-### 07/23/2018 - 1.0.0
+## v1.0.1 (07/24/2018)
+
+### Added
+* Added `brand()` in `Product` object.
+* Added `listPrice()` in `Product` object.
+* Added `getBrands()` in `Results` object.
+
+
+## v1.0.0 (07/23/2018)
 * Version 1.0 - Initial Release